% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmla_funcs.R
\name{fmla_from_dframe}
\alias{fmla_from_dframe}
\title{Create a formula from column names of a df}
\usage{
fmla_from_dframe(y, dframe)
}
\arguments{
\item{y}{A character vector with one element. The name of the response variable.}

\item{dframe}{A data frame like object with the named columns to use in the formula.}
}
\value{
A formula.
}
\description{
Create a formula from column names of a df
}
\examples{
dframe <- data.frame(y=c(1, 1, -1), x1=c("a", "b", "c"), x2=runif(3), x3=rnorm(3))
fmla_from_names("y", dframe)
}
