% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmla_funcs.R
\name{fmla_from_names}
\alias{fmla_from_names}
\title{Create a formula from names}
\usage{
fmla_from_names(y, nims)
}
\arguments{
\item{y}{A character vector with one element. The name of the response variable.}

\item{nims}{A character vector of any length. The names of the independent variables or features.}
}
\value{
A formula.
}
\description{
Create a formula from names
}
\examples{
nims <- c("x1", "x2", "x3")
fmla_from_names("y", nims)
}
